Refactor scanner into submod.
